pyproject.toml
[project]
name = "agentic-patterns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "langchain-groq>=1.1.1",
    "google-cloud-aiplatform>=1.132.0",
    "langchain-google-genai>=4.1.2",
    "google-adk>=1.21.0",
    "deprecated>=1.3.1",
    "pydantic>=2.12.5",
    "nest-asyncio>=1.6.0",
    "crewai>=1.6.1",
    "litellm>=1.80.11",
    "docling>=2.66.0",
    "langchain-docling>=2.0.0",
    "mcp>=1.25.0",
    "langchain-mcp-adapters>=0.2.1",
    "pymupdf>=1.26.7",
    "paddleocr>=3.0.0",
    "paddlepaddle>=3.0.0",
    "paddlex[ocr]>=3.0.0",
    "google-generativeai>=0.8.6",
    "langfuse>=2.0.0",
    "openinference-instrumentation-crewai>=0.1.0",
    "crewai-observe>=0.1.0",
    "a2a-sdk>=0.3.0",
    "httpx>=0.27.0",
    "uvicorn>=0.30.0",
]

# Note: After installing, run this to patch paddlex for langchain compatibility:
# sed -i '' 's/from langchain.docstore.document import Document/from langchain_core.documents import Document/g' .venv/lib/python*/site-packages/paddlex/inference/pipelines/components/retriever/base.py
# sed -i '' 's/from langchain.text_splitter import RecursiveCharacterTextSplitter/from langchain_text_splitters import RecursiveCharacterTextSplitter/g' .venv/lib/python*/site-packages/paddlex/inference/pipelines/components/retriever/base.py
